+++ src/commonMain/kotlin/dev/inmo/krontab/internal/CronDateTimeScheduler.kt
@@ -1,8 +1,8 @@
 package dev.inmo.krontab.internal
 
 import com.soywiz.klock.DateTime
-import dev.inmo.krontab.KronScheduler
-import dev.inmo.krontab.anyCronDateTime
+import dev.inmo.krontab.*
+import dev.inmo.krontab.collection.plus
 
 /**
  * Cron-oriented realisation of [KronScheduler]
@@ -13,40 +13,44 @@ import dev.inmo.krontab.anyCronDateTime
  * @see dev.inmo.krontab.EveryHourScheduler
  * @see dev.inmo.krontab.EveryDayOfMonthScheduler
  * @see dev.inmo.krontab.EveryMonthScheduler
+ * @see dev.inmo.krontab.EveryYearScheduler
  *
  * @see dev.inmo.krontab.builder.buildSchedule
  * @see dev.inmo.krontab.builder.SchedulerBuilder
  */
-@Deprecated("This class will get internal status in future")
-data class CronDateTimeScheduler internal constructor(
+internal data class CronDateTimeScheduler internal constructor(
     internal val cronDateTimes: List<CronDateTime>
 ) : KronScheduler {
     /**
-     * @return Near date using [cronDateTimes] list and getting the [Iterable.min] one
+     * @return Near date using [cronDateTimes] list and getting the [Iterable.minByOrNull] one
      *
      * @see toNearDateTime
      */
     override suspend fun next(relatively: DateTime): DateTime {
-        return cronDateTimes.map { it.toNearDateTime(relatively) }.minOrNull() ?: anyCronDateTime.toNearDateTime(relatively)
+        return cronDateTimes.mapNotNull { it.toNearDateTime(relatively) }.minOrNull() ?: getAnyNext(relatively)
     }
 }
 
-/**
- * @return New instance of [CronDateTimeScheduler] with all unique [CronDateTimeScheduler.cronDateTimes] of
- * [kronDateTimeSchedulers] included
- */
-@Suppress("NOTHING_TO_INLINE")
-fun merge(kronDateTimeSchedulers: List<CronDateTimeScheduler>) = CronDateTimeScheduler(
-    kronDateTimeSchedulers.flatMap { it.cronDateTimes }.distinct()
+internal fun mergeCronDateTimeSchedulers(schedulers: List<CronDateTimeScheduler>) = CronDateTimeScheduler(
+    schedulers.flatMap { it.cronDateTimes }
 )
 
 /**
- * @return Vararg shortcyut for [merge]
+ * @return New instance of [CronDateTimeScheduler] with all unique [CronDateTimeScheduler.cronDateTimes] of
+ * [kronSchedulers] included
  */
-@Suppress("NOTHING_TO_INLINE")
-inline fun merge(vararg kronDateTimeSchedulers: CronDateTimeScheduler) = merge(kronDateTimeSchedulers.toList())
+@Deprecated("Will be removed in next major release", ReplaceWith("merge", "dev.inmo.krontab"))
+fun merge(kronSchedulers: List<KronScheduler>) = kronSchedulers.apply { dev.inmo.krontab.merge() }
+
 /**
- * Use [merge] operation to internalcreate new [CronDateTimeScheduler] with all [CronDateTimeScheduler.cronDateTimes]
+ * @return Vararg shortcut for [dev.inmo.krontab.merge]
  */
 @Suppress("NOTHING_TO_INLINE")
-inline fun CronDateTimeScheduler.plus(other: CronDateTimeScheduler) = merge(this, other)
+@Deprecated("Will be removed in next major release", ReplaceWith("merge", "dev.inmo.krontab"))
+inline fun merge(vararg kronDateTimeSchedulers: KronScheduler) = kronDateTimeSchedulers.apply { dev.inmo.krontab.merge() }
+/**
+ * @return Vararg shortcut for [dev.inmo.krontab.merge]
+ */
+@Suppress("NOTHING_TO_INLINE")
+@Deprecated("Will be removed in next major release", ReplaceWith("merge", "dev.inmo.krontab"))
+inline fun KronScheduler.plus(other: KronScheduler) = this + other